Update deployment build info
