--- pkg/deployments/tasks/20220906-batch-relayer-v4/bytecode/BalancerRelayer.json
@@ -1,3 +1,0 @@
-{
-  "creationCode": "60c060405234801561001057600080fd5b5060405161078738038061078783398101604081905261002f91610052565b60016000556001600160601b0319606092831b8116608052911b1660a0526100a3565b60008060408385031215610064578182fd5b825161006f8161008b565b60208401519092506100808161008b565b809150509250929050565b6001600160a01b03811681146100a057600080fd5b50565b60805160601c60a05160601c6106b36100d46000398060fc528061020f5250806058528061012052506106b36000f3fe6080604052600436106100385760003560e01c80637678922e146100885780638d928af8146100b3578063ac9650d8146100c857610083565b36610083576100813373ffffffffffffffffffffffffffffffffffffffff7f000000000000000000000000000000000000000000000000000000000000000016146102066100e8565b005b600080fd5b34801561009457600080fd5b5061009d6100fa565b6040516100aa919061050e565b60405180910390f35b3480156100bf57600080fd5b5061009d61011e565b6100db6100d6366004610480565b610142565b6040516100aa919061052f565b816100f6576100f68161026e565b5050565b7f000000000000000000000000000000000000000000000000000000000000000090565b7f000000000000000000000000000000000000000000000000000000000000000090565b606061014c61029b565b8167ffffffffffffffff8111801561016357600080fd5b5060405190808252806020026020018201604052801561019757816020015b60608152602001906001900390816101825790505b50905060005b82811015610257576102388484838181106101b457fe5b90506020028101906101c691906105e3565b8080601f0160208091040260200160405190810160405280939291908181526020018383808284376000920191909152505073ffffffffffffffffffffffffffffffffffffffff7f000000000000000000000000000000000000000000000000000000000000000016929150506102b4565b82828151811061024457fe5b602090810291909101015260010161019d565b50610260610337565b610268610348565b92915050565b610298817f42414c000000000000000000000000000000000000000000000000000000000061034f565b50565b6102ad600260005414156101906100e8565b6002600055565b6060600060608473ffffffffffffffffffffffffffffffffffffffff16846040516102df91906104ef565b600060405180830381855af49150503d806000811461031a576040519150601f19603f3d011682016040523d82523d6000602084013e61031f565b606091505b509150915061032e82826103ca565b95945050505050565b4780156102985761029833826103f4565b6001600055565b7f08c379a000000000000000000000000000000000000000000000000000000000600090815260206004526007602452600a808404818106603090810160081b958390069590950190829004918206850160101b01602363ffffff0060e086901c160160181b0190930160c81b60445260e882901c90606490fd5b606082156103d9575080610268565b8151156103e95781518083602001fd5b6102686101ae61026e565b610403814710156101a36100e8565b60008273ffffffffffffffffffffffffffffffffffffffff16826040516104299061050b565b60006040518083038185875af1925050503d8060008114610466576040519150601f19603f3d011682016040523d82523d6000602084013e61046b565b606091505b5050905061047b816101a46100e8565b505050565b60008060208385031215610492578182fd5b823567ffffffffffffffff808211156104a9578384fd5b818501915085601f8301126104bc578384fd5b8135818111156104ca578485fd5b86602080830285010111156104dd578485fd5b60209290920196919550909350505050565b6000825161050181846020870161064d565b9190910192915050565b90565b73ffffffffffffffffffffffffffffffffffffffff91909116815260200190565b6000602080830181845280855180835260408601915060408482028701019250838701855b828110156105d6577fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffc088860301845281518051808752610599818989018a850161064d565b601f017fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe01695909501860194509285019290850190600101610554565b5092979650505050505050565b60008083357fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe1843603018112610617578283fd5b83018035915067ffffffffffffffff821115610631578283fd5b60200191503681900382131561064657600080fd5b9250929050565b60005b83811015610668578181015183820152602001610650565b83811115610677576000848401525b5050505056fea2646970667358221220063fdaaad6c7f32fb6dd53f100269a9ea16c7241c4a9587b1f83f82a35b5424264736f6c63430007010033"
-}
