--- pkg/deployments/tasks/20220906-batch-relayer-v4/output/goerli.json
@@ -1,5 +1,0 @@
-{
-  "BatchRelayerLibrary": "0x992b827a42ac301DF16Fedb3eBbEA59AC1DC35DE",
-  "timestamp": 1663176235248,
-  "BalancerRelayer": "0x2Fb8aD26faDC971464A17407e9213D1A22befc8a"
-}
