+++ pkg/deployments/tasks/20220916-batch-relayer-v4/readme.md
@@ -1,0 +1,10 @@
+# 2022-09-16 - Batch Relayer V4
+
+Deployment of the fourth `BalancerRelayer` using `BatchRelayerLibrary`, for combining multiple operations (swaps, joins, etc.) in a single transaction.
+
+This new version provides quality of life improvements, like read-only chained references and ways to read them.
+
+## Useful Files
+
+- [`BalancerRelayer` ABI](./abi/BalancerRelayer.json)
+- [`BatchRelayerLibrary` ABI](./abi/BatchRelayerLibrary.json)
