+++ pkg/deployments/tasks/deprecated/20220720-batch-relayer-v3/readme.md
@@ -2,7 +2,7 @@
 
 > ⚠️ **DEPRECATED: do not use** ⚠️
 >
-> This relayer has been deprecated in favor of an [updated version](../../20220906-batch-relayer-v4) with new features, such as read-only references and a 'peek' reference public method.
+> This relayer has been deprecated in favor of an [updated version](../../20220916-batch-relayer-v4) with new features, such as read-only references and a 'peek' reference public method.
 
 Deployment of the second `BalancerRelayer` using `BatchRelayerLibrary`, for combining multiple operations (swaps, joins, etc.) in a single transaction.
